@@ -8,7 +8,7 @@ plugins {
 group 'de.unijena.cheminf'
 version '1.0.1.0'
 
-sourceCompatibility = 1.11
+sourceCompatibility = 1.17
 
 //Creates javadoc and sources jars
 java {
